+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-# v0.1.0
+# 0.2.0
+* Asses are now precompiled.
+* Changed `resizeSpeed` for `resizeInterval`.
 
+# v0.1.0
 * Added `resizeSpeed` prop.